Move builtin utility types to yash-core
