+++ yash-core/src/lib.rs
@@ -17,6 +17,7 @@
 //! TODO Elaborate
 
 pub mod alias;
+pub mod builtin;
 pub mod env;
 pub mod exec;
 pub mod expansion;
